clean up

+++ processor/src/main/java/my/raptive/url/processor/repository/ShardCount.java
@@ -15,5 +15,5 @@ import lombok.experimental.Accessors;
 public class ShardCount {
 
     @Id
-    private long shardCount;
+    private int shardCount;
 }
